move attack logic to EnemyController.cs
