+++ Assets/Prefabs/Characters/Enemy 1.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 22dfa69fcf8342148bed61d6eee040eb
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
