+++ Assets/Scenes/Samples/EnemyAttacks.unity
@@ -7863,6 +7863,23 @@ Rigidbody2D:
   m_SleepingMode: 1
   m_CollisionDetection: 0
   m_Constraints: 0
+--- !u!1 &1090854094 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 5580964773805009162, guid: 5de27c63050c7b84099a93bfd18aa2bb, type: 3}
+  m_PrefabInstance: {fileID: 1287376676106477322}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &1090854100
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1090854094}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c999010e02a31ff4db4a1571703ee32e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &1221295313
 GameObject:
   m_ObjectHideFlags: 0
@@ -8357,6 +8374,10 @@ PrefabInstance:
     serializedVersion: 3
     m_TransformParent: {fileID: 0}
     m_Modifications:
+    - target: {fileID: 163350721441205344, guid: 5de27c63050c7b84099a93bfd18aa2bb, type: 3}
+      propertyPath: m_Layer
+      value: 7
+      objectReference: {fileID: 0}
     - target: {fileID: 202247549049000719, guid: 5de27c63050c7b84099a93bfd18aa2bb, type: 3}
       propertyPath: m_LocalPosition.x
       value: 17.08
@@ -8397,14 +8418,49 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
+    - target: {fileID: 507465705623202848, guid: 5de27c63050c7b84099a93bfd18aa2bb, type: 3}
+      propertyPath: m_Layer
+      value: 7
+      objectReference: {fileID: 0}
+    - target: {fileID: 2018167636680774825, guid: 5de27c63050c7b84099a93bfd18aa2bb, type: 3}
+      propertyPath: m_Layer
+      value: 7
+      objectReference: {fileID: 0}
+    - target: {fileID: 3767186367271092461, guid: 5de27c63050c7b84099a93bfd18aa2bb, type: 3}
+      propertyPath: m_Layer
+      value: 7
+      objectReference: {fileID: 0}
+    - target: {fileID: 5350570460743989002, guid: 5de27c63050c7b84099a93bfd18aa2bb, type: 3}
+      propertyPath: m_Layer
+      value: 7
+      objectReference: {fileID: 0}
     - target: {fileID: 5580964773805009162, guid: 5de27c63050c7b84099a93bfd18aa2bb, type: 3}
       propertyPath: m_Name
       value: EnemyMelee
       objectReference: {fileID: 0}
+    - target: {fileID: 5580964773805009162, guid: 5de27c63050c7b84099a93bfd18aa2bb, type: 3}
+      propertyPath: m_Layer
+      value: 7
+      objectReference: {fileID: 0}
+    - target: {fileID: 5774138483983516601, guid: 5de27c63050c7b84099a93bfd18aa2bb, type: 3}
+      propertyPath: m_Layer
+      value: 7
+      objectReference: {fileID: 0}
+    - target: {fileID: 6015352937621917485, guid: 5de27c63050c7b84099a93bfd18aa2bb, type: 3}
+      propertyPath: m_Layer
+      value: 7
+      objectReference: {fileID: 0}
+    - target: {fileID: 8577031309454380906, guid: 5de27c63050c7b84099a93bfd18aa2bb, type: 3}
+      propertyPath: m_Layer
+      value: 7
+      objectReference: {fileID: 0}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
-    m_AddedComponents: []
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: 5580964773805009162, guid: 5de27c63050c7b84099a93bfd18aa2bb, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 1090854100}
   m_SourcePrefab: {fileID: 100100000, guid: 5de27c63050c7b84099a93bfd18aa2bb, type: 3}
 --- !u!1001 &8123176992794920350
 PrefabInstance:
@@ -8417,6 +8473,10 @@ PrefabInstance:
     - target: {fileID: 1512645617981690473, guid: 3e7dde9e87206674f8f4d83b42ee87a4, type: 3}
       propertyPath: m_Name
       value: EnemyRanged
+      objectReference: {fileID: 0}
+    - target: {fileID: 2547346709570462281, guid: 3e7dde9e87206674f8f4d83b42ee87a4, type: 3}
+      propertyPath: attackRange
+      value: 2
       objectReference: {fileID: 0}
     - target: {fileID: 8096218100924724951, guid: 3e7dde9e87206674f8f4d83b42ee87a4, type: 3}
       propertyPath: m_LocalPosition.x
