+++ Assets/Scripts/Enemy Attacks/ProjectileScript.cs
@@ -7,20 +7,25 @@ public class ProjectileScript : MonoBehaviour
 
     [SerializeField] int damage;
     public float speed;
+    public AudioClip[] arrowSounds;
 
     private float timer;
     private Rigidbody2D rb;
     private GameObject player;
     private PlayerStats playerStats;
 
+    private AudioSource audioSource;
+
     void Start()
     {
+        audioSource = GetComponent<AudioSource>();
         rb = GetComponent<Rigidbody2D>();
         player = GameObject.FindWithTag("Player");
         playerStats = player.GetComponent<PlayerStats>();
 
         Vector3 direction = player.transform.position - transform.position;
         rb.velocity = new Vector2(direction.x, direction.y).normalized * speed;
+        PlayRandomArrowSound();
     }
 
     void Update()
@@ -37,6 +42,15 @@ public class ProjectileScript : MonoBehaviour
         if (other.CompareTag("PlayerBody"))
         {
             playerStats.TakeDamage(damage);
+        }
+    }
+
+    private void PlayRandomArrowSound()
+    {
+        if (arrowSounds.Length > 0)
+        {
+            AudioClip selectedSound = arrowSounds[Random.Range(0, arrowSounds.Length)];
+            audioSource.PlayOneShot(selectedSound);
         }
     }
 }
